Properly organize ELF functionality
The ELF logic is distributed all over the place, with no organization
principle in mind. Put it into a proper module structure.
