+++ src/elf/mod.rs
@@ -1,0 +1,8 @@
+mod cache;
+mod parser;
+mod resolver;
+mod types;
+
+pub use cache::ElfCache;
+pub use parser::Elf64Parser;
+pub use resolver::ElfResolver;
